refactor(captain): extract AI command registration into rootcmd
Move AI runtime command registration into the internal root command package to keep CLI command wiring modular without changing behavior.
